Add yaN and yaN-luk data from vidyut-prakriya
Generated from a dev build of `vidyut-prakriya` [1], a Paninian word
generator.

Our generation setup is not checked in yet but will be within a week of
this commit. The setup will be available in the `examples` directory of
the `vidyut-prakriya` project.

Sample usage, for future reference. Note that this API is not stable and
might change in the future.

```
cargo run --release --example create_tinantas_babylon -- \
--sanadi yaN \
--desc "vidyud-yantreRa janitAni yaNanteByas tiNantAni" > vidyut_yaN.babylon
```

```
cargo run --release --example create_tinantas_babylon -- \
--sanadi yaNLuk \
--desc "vidyud-yantreRa janitAni yaNluganteByas tiNantAni" > vidyut-yaN-luk.babylon
```

[1]: https://github.com/ambuda-org/vidyut/tree/main/vidyut-prakriya